+++ content/tabkit.js
@@ -4868,13 +4868,7 @@ window.tabkit = new function _tabkit() { // Primarily just a 'namespace' to hide
   this.initListeners.push(this.initTabbarPosition);
 
   this.TabBar = this.TabBar || {};
-  this.TabBar.Callbacks = this.TabBar.Callbacks || {}
-  this.TabBar.Callbacks.initTabbarPositionOnPostInit = function initTabbarPositionOnPostInit(event) {
-    // HACK: Run again to workaround a strange bug that
-    // The splitter is not showing properly on OSX on startup
-    tk.moveTabbar();
-  }
-  this.postInitListeners.push(this.TabBar.Callbacks.initTabbarPositionOnPostInit);
+  this.TabBar.Callbacks = this.TabBar.Callbacks || {};
 
   /// Event listeners:
   this.bug608589workaround = function bug608589workaround(event) {