+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Bug Fix
 - Fix "Splitter cannot be dragged when vertical tabbar has a small width" (#65)
 - Fix "Vertical Tab Bar Has Horizontal Tabs Jammed In" (#58) by removing a "fix" that is no longer needed by OSX
+- No longer remove the original tab when copying a single tab to a different browser window
 
 
 ## 0.11.2